simFun.m
changed formatting of comments so that running "occur" (in emacs) with
argument "^[[:space]]*%% " shows the overall logical structure of the
code (sort of).

commented out the "bench.alpha = 0" on line 228.  this was set for
convenience way back when i was asked to run ETS imported data.  The
same effect can be had by calling ...,'alpha',0... in the simFun
argument list.