detectCollisionsSparse
changed collision detection from point-to-line-segment to
point-to-points, checking fiber-fiber and fiber-elbow interferences.
This violates the ideal geometry by 14.5 microns in the center of the
arm, but the surface tolerance is 25 microns.  There is a small region
in which collisions will occur, but not be detected by the current
method.

on the other hand, it makes the dynamics of fiber-arm collision more
accurate in most cases.