generateTrajectory2
change to use Fast map for initial move

in some cases when the initial position is at the positive-going home,
floating point roundoff causes the calculated strtBin.thtP to be < 0,
usually by ~1e-13 or something insignificant.  Values < 0 are set to
zero with some text going to stdout to document the situation.

Properly, the system should fail if the value is "large" and negative,
but that's not coded in yet.