loadCfgXml
cleaned up indentation an logic:

If there is only one xml file in the current directory, then use
that one.

If there are multiple files, then use "usedXMLFile.xml" if it is in
the directory specified (or the current directory).

If there are multiple xml files, none of which are named
"usedXMLFile.xml", then bring up the dialog box.

NOTE: To avoid "usedXMLFile.xml", the cfgFile name MUST be specified,
even if is a junk name