generateTrajectory2.m
cleaned up code a little.  found that output.nmax is not really used
anywhere.  Only in realizeTrajectory2.m, and there only to get the
length of the vector.

added some comment lines on output.lmax.  this is a record-keeping
variable that gives the maximum length of a vector in ThtP, ThtN or
Phi.